pyproject.toml
[tool.poetry]
name = "remnote-graph-rag"
version = "0.1.0"
description = "A Graph RAG system based on RemNote knowledge base."
authors = ["Fuku Ro <keguryanova@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.14"
pydantic-settings = "^2.12.0"
ollama = "^0.6.1"
llama-index = "^0.14.12"
llama-index-storage-docstore-redis = "^0.4.1"
llama-index-storage-index-store-redis = "^0.5.1"
llama-index-embeddings-huggingface = "^0.6.1"
llama-index-vector-stores-redis = "^0.6.2"
llama-index-graph-stores-neo4j = "^0.5.1"
llama-index-llms-ollama = "^0.9.1"
llama-index-embeddings-ollama = "^0.8.6"
llama-index-postprocessor-cohere-rerank = "^0.5.1"
langchain = "^1.2.0"
langchain-ollama = "^1.0.1"
tavily-python = "^0.7.17"
reflex = "^0.8.26"
plotly = "^6.5.2"
pymupdf = "^1.26.7"
paddleocr = { extras = ["all"], version = "^3.3.2" }
paddlepaddle = "^3.2.2" # To use GPU paddlepaddle-gpu need to be installed
neo4j = "5.28.2"


[tool.poetry.group.test.dependencies]
pyvis = "^0.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
